+++ tools/src/tools/runner.cpp
@@ -9,7 +9,7 @@ namespace {
 [[nodiscard]] auto load_or_create_state(App const& app) {
 	static auto const logger = bave::Logger{"BaveTools"};
 	auto ret = std::make_shared<State>();
-	ret->assets_path = app.get_assets_path();
+	ret->assets_path = app.get_assets_dir();
 	if (ret->load()) {
 		logger.info("loaded State from '{}'", State::path_v.as_view());
 	} else {